requirements.txt
streamlit
PyPDF2
langchain
faiss-cpu
openai
python-dotenv
langchain_community
